Add assert in classes to document key assumptions
There are several methods in TaskList that require a valid
TaskList to be passed in for the methods to work as intended.
Adding a check to ensure the taskList is not null will help
with this.

Similarly, there are methods in UI.java and SaveFileManager
where the parameter passed in should be non-null. Checks
have been added for those methods as well.

Additionally, build.gradle has been modified to enable assertions.