Refactor Parser.java to reduce method length
Parser.java only had one method, with each command having its
own pattern.matcher line, causing the method to be very long.

To solve this, a CommandPattern class was added to abstract out
the patterns. By iterating through a list of CommandPatterns,
the logic of the code is much cleaner, clearer and shorter.

A new command, help, is also added. It will provide the user with
a list of possible commands, and show the proper input format of
each command.